apps/api/requirements.txt
# VisionSathi API Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# ML/AI
torch==2.1.2
transformers==4.37.0
Pillow==10.2.0

# Moondream
# Note: Install from source or pip when available
# moondream==0.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Caching
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.0
isort==5.13.2
mypy==1.8.0
